refactor tauri workspace browser structure
